requirements.txt
streamlit
opencv-python
numpy
torch
torchvision
Pillow
mediapipe
